refactor: gender enum constants로 이동
